Seperate database connection and server connection
