+++ database/db.js
@@ -1,0 +1,10 @@
+const mysql = require('mysql');
+
+const connection = mysql.createConnection({
+  host: 'localhost',
+  user: 'root',
+  password: '',
+  database: 'listing_body'
+});
+
+module.exports = connection;
